refactor(ae2/pattern-buffer): 拆分镜像位配置器的客户端逻辑
将 PatternBufferProxyPositionConfigurator 中的高亮、传送按钮等
纯客户端操作提取到新的 PatternBufferProxyPositionConfiguratorClient，
使配置器本体不再引用 Minecraft / BlockHighlightHandler 等客户端类。
